added notation actions
