+++ framework/actions/CMakeLists.txt
@@ -24,8 +24,6 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/actionsmodule.h
     ${CMAKE_CURRENT_LIST_DIR}/iactionsdispatcher.h
     ${CMAKE_CURRENT_LIST_DIR}/action.h
-    ${CMAKE_CURRENT_LIST_DIR}/actions.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/actions.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/actionsdispatcher.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/actionsdispatcher.h
     )
