+++ mu4/domain/notation/CMakeLists.txt
@@ -22,13 +22,15 @@ set(MODULE notation)
 find_package(Qt5 COMPONENTS Svg REQUIRED)
 
 set(MODULE_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/notationdomainmodule.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/notationdomainmodule.h
     ${CMAKE_CURRENT_LIST_DIR}/inotation.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationcreator.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationinputstate.h
     ${CMAKE_CURRENT_LIST_DIR}/inotationselection.h
     ${CMAKE_CURRENT_LIST_DIR}/notationtypes.h
-    ${CMAKE_CURRENT_LIST_DIR}/notationdomainmodule.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/notationdomainmodule.h
+    ${CMAKE_CURRENT_LIST_DIR}/notationactions.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/notationactions.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/notation.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/notation.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/notationcreator.cpp
