+++ mu4/scenes/notation/CMakeLists.txt
@@ -36,7 +36,7 @@ set(MODULE_SRC
     )
 
 set(MODULE_LINK
-    actions
+    notation
     )
 
 include(${PROJECT_SOURCE_DIR}/build/module.cmake)
